refactor: use shared supabase client
