+++ services/database.ts
@@ -1,4 +1,4 @@
-import { supabase } from './supabase';
+import { supabase } from '../lib/supabase';
 import { 
   Product, 
   Category,
